chore: move Talk and user file to component folder
